+++ src/rafimage.cpp
@@ -51,7 +51,7 @@ namespace Exiv2 {
         return "image/x-fuji-raf";
     }
 
-    int RafImage::pixelWidth() const
+    int64_t RafImage::pixelWidth() const
     {
         auto widthIter = exifData_.findKey(Exiv2::ExifKey("Exif.Photo.PixelXDimension"));
         if (widthIter != exifData_.end() && widthIter->count() > 0) {
@@ -60,7 +60,7 @@ namespace Exiv2 {
         return 0;
     }
 
-    int RafImage::pixelHeight() const
+    int64_t RafImage::pixelHeight() const
     {
         auto heightIter = exifData_.findKey(Exiv2::ExifKey("Exif.Photo.PixelYDimension"));
         if (heightIter != exifData_.end() && heightIter->count() > 0) {
